% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/abstract.r
\name{abstract}
\alias{abstract}
\title{Abstract an R Object}
\usage{
abstract(x)
}
\arguments{
\item{x}{An R object}
}
\description{
A simple function to "abstract" an R object into the R code which would
produce it.  A simple wrapper around \code{base::dump()}.
}
\examples{
df <- data.frame(index=1:5, x=rnorm(5))
df
abstract(df)
}

