% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bfs.r
\name{find.bfs}
\alias{find.bfs}
\title{Find Basic Feasible Solutions}
\usage{
find.bfs(A, b)
}
\arguments{
\item{A}{Coefficient matrix of the convex set.}

\item{b}{Right hand side.}
}
\value{
A 'bfs' object.
}
\description{
Find Basic Feasible Solutions
}

