% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bfs.r
\name{find.ep}
\alias{find.ep}
\title{Find Extreme Points}
\usage{
find.ep(bfs)
}
\arguments{
\item{bfs}{A bfs object, returned from \code{find.bfs()}.}
}
\value{
An 'ep' class object.
}
\description{
Find extreme points.
}

