% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bins.R
\name{gen_bin}
\alias{gen_bin}
\title{Automatically adds binned variables to data frame}
\usage{
gen_bin(df, x, binvec)
}
\arguments{
\item{df}{Data frame to be transformed.}

\item{x}{Variable to be classified in bins.}

\item{binvec}{Vector with increasing sequence to create bins.
(e.g., if binvec=c(1,2,3) two bins are created from intervals (1,2] and (2,3]).}
}
\value{
Original data frame but with added bins.
}
\description{
Automatically adds binned variables to data frame
}
