% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumulative.R
\name{gen_cumulative}
\alias{gen_cumulative}
\title{Automatically adds variable with cumulative change after episodes to data frame}
\usage{
gen_cumulative(df, delta, dates, kvec, id, time)
}
\arguments{
\item{df}{Data frame to be transformed.}

\item{delta}{Vector with changes.}

\item{dates}{Dates with episodes (equals 1 if episode occurred).}

\item{kvec}{(Possibly of a vector of) steps to compute cumulative changes
(e.g., if k=2 it computes delta(t)+delta(t+1)+delta(t+2) for each episode).}

\item{id}{Individual dimension of panel (if not panel data provide constant variable).}

\item{time}{Time dimension of panel.}
}
\value{
Same data frame initially provided but with added cumulative change variables.
}
\description{
Automatically adds variable with cumulative change after episodes to data frame
}
