% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_projections.R
\name{gen_data_lp}
\alias{gen_data_lp}
\title{Prepare data frame to run local projection}
\usage{
gen_data_lp(df, endog, exogl = NULL, lags, horizon, id, time, growth = TRUE)
}
\arguments{
\item{df}{Data frame that will be transformed.}

\item{endog}{String with endogenous variable. It should represent change between t and t-1 if growth = TRUE and level if growth = FALSE.
See explanation in growth parameter below for details.}

\item{exogl}{Vector with exogenous variables of LP that are lagged, as strings (default is NULL).}

\item{lags}{Number of lags used in LP.}

\item{horizon}{Horizon used in LP.}

\item{id}{Individual dimension of panel, as string (provide constant variable if data is not panel).}

\item{time}{Time dimension of the data, as string.}

\item{growth}{If LHS of regression is growth rate of variable of interest y between t and t+k set equal to TRUE (default).
If it is level set equal to FALSE. Therefore:
\itemize{
\item If growth=TRUE, endogenous variables on the LHS are differences between: (t-1) and t, t(-1) and (t+1),..., (t-1) and (t+k). In this case, endog should
represent a vector with first differences of variable of interest y.
\item If growth=FALSE, endogenous variables are y(t), y(t+1),...,y(t+k). In this case endog should be the level of the variable of interest y.
}}
}
\value{
Data frame provided initially, but with added variables.
}
\description{
This function prepares a data frame to run local projections.
It is not necessary to use this function before running a local projection, except when the argument gen_data is set to false in the function
local_projection(). It may be useful to use this function if one runs many local projections with the same variables.
}
