% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_projections.R
\name{gen_forward_delta}
\alias{gen_forward_delta}
\title{Automatically adds change of variable between t and t+k to data frame}
\usage{
gen_forward_delta(df, delta, horizon, id, time)
}
\arguments{
\item{df}{Data frame that will be transformed.}

\item{delta}{Vector of changes in a variable of interest x (that is, x(t)-x(t-1)).}

\item{horizon}{Maximum horizon to compute x(t+k)-x(t-1).}

\item{id}{Individual dimension of panel (provide constant variable if data is not panel).}

\item{time}{Time dimension of the data.}
}
\value{
Data frame provided initially, but with the added variables.
}
\description{
Automatically adds change of variable between t and t+k to data frame
}
