% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lags_and_leads.R
\name{gen_lead}
\alias{gen_lead}
\title{Automatically adds leads to data frame}
\usage{
gen_lead(df, x, lead, id, time)
}
\arguments{
\item{df}{Data frame to be transformed.}

\item{x}{Name of the variable used to compute the leads.}

\item{lead}{Maximum lead used (e.g.,lead = 2 creates variable x(t+1) and x(t+2)).}

\item{id}{Individual dimension of panel data (if data is not a panel a constant variable should be provided).}

\item{time}{Time dimension of the data.}
}
\value{
Provided data frame df with the added lagged variables.
}
\description{
This function takes a data frame and returns the same data frame with the additional variables of leads.
}
