% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bins.R
\name{get_bin}
\alias{get_bin}
\title{Classifies variables according to bins}
\usage{
get_bin(x, inf, sup)
}
\arguments{
\item{x}{Variable to be classified.}

\item{inf}{Lower bound of interval.}

\item{sup}{Upper bound of interval.}
}
\value{
Dummy that equals 1 if variable belongs to bin.
}
\description{
Returns 1 if x> inf and x<=sup and zero otherwise (and NA if the variable is NA).
}
