% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/levels_and_differences.R
\name{get_dif}
\alias{get_dif}
\title{First difference of a vector with NA in first row}
\usage{
get_dif(x)
}
\arguments{
\item{x}{Variable to compute first difference.}
}
\value{
First difference with NA in first row.
}
\description{
First difference of a vector with NA in first row
}
