% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/levels_and_differences.R
\name{get_fake_level}
\alias{get_fake_level}
\title{Gets the "fake" level of some data from changes,}
\usage{
get_fake_level(delta)
}
\arguments{
\item{delta}{Vector with changes of data to compute fake level.}
}
\value{
Vector with fake levels (only useful to back out first differences).
}
\description{
This function computes the level of some data from changes. We say it is the "fake" level
because if there is missing data (holes in the time series), the function still computes some kind of level.
In particular, after each hole in the series, a 0 is assigned to the level variable.
If one uses the output of this function to compute first differences, we get the right first differences (other than that, the output has no meaning).
This function is useful to use in packages that requires some variable in level as input, but that compute first differences before using it.
}
