% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_projections.R
\name{get_forward_delta}
\alias{get_forward_delta}
\title{Computes change between (t-1) and t+k from a vector of changes delta between t-1 and t}
\usage{
get_forward_delta(delta, k)
}
\arguments{
\item{delta}{Vector with changes between t and (t-1) of variable.}

\item{k}{Steps forward,}
}
\value{
Change between (t+k) and (t-1) of variable.
}
\description{
Computes change between (t-1) and t+k from a vector of changes delta between t-1 and t
}
