% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lags_and_leads.R
\name{get_lag}
\alias{get_lag}
\title{Computes lag of vector}
\usage{
get_lag(x, k)
}
\arguments{
\item{x}{Vector used to compute lags.}

\item{k}{Number of lags.}
}
\value{
Vector x(t-k), with NA in the first k entries.
}
\description{
Computes lag of vector
}
