% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lags_and_leads.R
\name{get_lead}
\alias{get_lead}
\title{Computes lead of vector}
\usage{
get_lead(x, k)
}
\arguments{
\item{x}{Vector used to compute leads.}

\item{k}{Number of leads.}
}
\value{
Vector x(t+k), with NA in the last k entries.
}
\description{
Computes lead of vector
}
