% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/levels_and_differences.R
\name{get_level}
\alias{get_level}
\title{Gets the level of some data from changes}
\usage{
get_level(delta, dates, ref)
}
\arguments{
\item{delta}{Vector of changes in a variable of interest x (that is x(t)-x(t-1)).}

\item{dates}{Vector of dates associated to each value of x (it must be in ascending order).}

\item{ref}{Date at which we normalize the level to zero.}
}
\value{
Vector with normalized level.
}
\description{
Gets the level of some data from changes
}
