% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/levels_and_differences.R
\name{get_logdif}
\alias{get_logdif}
\title{Log difference of a vector with NA in first row}
\usage{
get_logdif(x)
}
\arguments{
\item{x}{Variable to compute log differences.}
}
\value{
Log difference with NA in first row.
}
\description{
Log difference of a vector with NA in first row
}
