% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/display.r
\name{latex.bfs}
\alias{latex.bfs}
\title{Display Basic Feasible Solutions}
\usage{
\method{latex}{bfs}(object, digits = 3)
}
\arguments{
\item{object}{A 'bfs' object.}

\item{digits}{Number of rounding digits.}
}
\description{
Display a bfs in LaTeX.  For use with knitr/sweave.
}

