% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/latex.r
\name{latex.ep}
\alias{latex.ep}
\title{Display Extreme Points}
\usage{
\method{latex}{ep}(object, digits = 3)
}
\arguments{
\item{object}{An 'ep' object.}

\item{digits}{Number of rounding digits.}
}
\description{
Display Extreme Points
}

