% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/latex.r
\name{latex}
\alias{latex}
\alias{latex.bfs}
\alias{latex.ep}
\alias{latex.matrix}
\alias{latex.numeric}
\title{LaTeX printing}
\usage{
latex(object, ...)

\method{latex}{matrix}(object, digits = 3, asTranspose = FALSE, ...)

\method{latex}{numeric}(object, digits = 3, asTranspose = FALSE, ...)

\method{latex}{bfs}(object, digits = 3, ...)

\method{latex}{ep}(object, digits = 3, ...)
}
\arguments{
\item{object}{Object to print in LaTeX.}

\item{...}{Additional arguments.}

\item{digits}{Number of rounding digits}

\item{asTranspose}{Display as the transpose?}
}
\description{
Displaying objects of the appropriate type, typeset in
LaTeX.  For use with knitr/sweave.
}
\examples{
x <- matrix(1:25, 5)
latex(x)
}

