% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_projections.R
\name{local_projection}
\alias{local_projection}
\title{Runs local projection}
\usage{
local_projection(
  df,
  endog,
  exogc = NULL,
  exogl = NULL,
  shock,
  lags = 0,
  horizon,
  method,
  append_formula = NULL,
  id,
  time,
  conf_level = 0.9,
  gen_data = TRUE,
  growth = TRUE,
  ...
)
}
\arguments{
\item{df}{Data frame with variables used.}

\item{endog}{String with endogenous variable. It should represent change between t and t-1 if growth = TRUE and level if growth = FALSE.
See explanation in growth parameter below for details.}

\item{exogl}{Vector with exogenous variables that are lagged, as strings (default is NULL).}

\item{shock}{Shock variable used to report IRF (as string).}

\item{lags}{Number of lags used for variables in exogl.}

\item{horizon}{Maximum horizon used to compute IRFs.}

\item{method}{Function used to run regressions (e.g., lm, felm, etc). If function requires package it should be loaded.}

\item{append_formula}{string to append to formula of regression.
By default formula is (endogenous variable ~ contemporaneous controls + exogenous controls). For instance, if one uses method=felm and want
to add country fixed effects, one should use append_formula="| country". Similarly, one can add instrumental variables that way.}

\item{id}{Individual dimension of panel, as string (provide constant variable if data is not panel).}

\item{time}{Time dimension of the data, as string.}

\item{conf_level}{Confidence level for confidence bands (default is 10\%).}

\item{gen_data}{Use TRUE if data frame was not prepared before hand using gen_data_lp().}

\item{growth}{If LHS of regression is growth rate of variable of interest y between t and t+k set equal to TRUE (default).
If it is level set equal to FALSE. Therefore:
\itemize{
\item If growth=TRUE, endogenous variables on the LHS are differences between: (t-1) and t, t(-1) and (t+1),..., (t-1) and (t+k). In this case, endog should
represent a vector with first differences of variable of interest y.
\item If growth=FALSE, endogenous variables are y(t), y(t+1),...,y(t+k). In this case endog should be the level of the variable of interest y.
}}

\item{...}{Additional arguments that will be passed to regression method chosen.}
}
\value{
A list with:
\itemize{
\item Impulse response functions (irf);
\item Lower confidence bands (conf_low);
\item Upper confidence bands (conf_high)
\item Vector with horizons 0:horizon (hvec);
\item List with regressions output (e.g., to access regression of horizon k call output$regressions[[\link{k}]]).
}
}
\description{
This function runs a local projection.
}
