% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lptools-package.r
\docType{package}
\name{lptools-package}
\alias{lptools-package}
\title{lptools:  Linear programming utilities}
\description{
TODO
}
\author{
Drew Schmidt \email{wrathematics AT gmail.com}
}
\keyword{Package}

