% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/matdisp.r
\name{matdisp}
\alias{matdisp}
\title{Display a Matrix in LaTeX}
\usage{
matdisp(A)
}
\arguments{
\item{A}{Matrix to display.}
}
\value{
A character string of the matrix data inside an \code{align*}
environment.
}
\description{
Display a Matrix in LaTeX
}
\examples{
x <- matrix(1:25, 5)
matdisp(x)
}

