% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bfs.r
\name{optimize}
\alias{maximize}
\alias{minimize}
\alias{optimize}
\title{Optimize Linear Programming Problem}
\usage{
minimize(z, ep)

maximize(z, ep)
}
\arguments{
\item{z}{The coefficients of the objective function, including zeros from
slack variables if applicable.}

\item{ep}{A set of extreme points (the output of \code{find.ep()}).}
}
\description{
Given a set of extreme points and the coefficients of an
objective function
}

