% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/transform.r
\name{slack}
\alias{slack}
\alias{slackdir}
\title{Add Slack Variables}
\usage{
slack(A, rows)

slackdir(A, rows)
}
\arguments{
\item{A}{Coefficient matrix.}

\item{rows}{Index of rows which are _not_ equality constraints.  If left blank,
all are assumed to be inequality constraints.}
}
\description{
From a coefficient matrix, add slack variables.
}
\details{
This operation comes in two flavors, goverened by functions
\code{slack()} and \code{slackdir()}.  For the former, given a
coefficient matrix \code{A} and an index of rows \code{rows},
this will return the coefficient matrix with slack variables
added appropriately.  The latter does the same, but for
extreme direction problems; in particular, it adds the constraint
for the sum of the \code{d}'s.
}

