% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simplex.r
\name{tableau}
\alias{tableau}
\title{Print a Simplex Table}
\usage{
tableau(x, Bs, digits = 0)
}
\arguments{
\item{x}{A matrix of values for the simplex table.}

\item{Bs}{The indices of the current basis.}

\item{digits}{The number of decimal digits to use in the table.}
}
\description{
Using the notation of Bazaraa et. al., given a table \code{x} and a
basis vector \code{Bs}, this will produce the LaTeX for a simplex
tableau.
}

