% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/display.r
\name{latex.matrix}
\alias{latex.matrix}
\title{Display a Matrix}
\usage{
\method{latex}{matrix}(object, digits = 3)
}
\arguments{
\item{digits}{Number of rounding digits.}

\item{A}{Matrix to display.}
}
\description{
Display a matrix in LaTeX.  For use with knitr/sweave.
}
\examples{
x <- matrix(1:25, 5)
display(x)
}

