% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/latex.r
\name{latex}
\alias{latex}
\alias{latex.bfs}
\alias{latex.ep}
\alias{latex.lp_solution}
\alias{latex.matrix}
\alias{latex.numeric}
\title{LaTeX printing}
\usage{
latex(object, ...)

\method{latex}{matrix}(object, inline = FALSE, digits = 3,
  showName = TRUE, asTranspose = FALSE, system = FALSE, ..., dispname)

\method{latex}{numeric}(object, inline = FALSE, digits = 3,
  showName = TRUE, asTranspose = FALSE, ...)

\method{latex}{bfs}(object, inline = FALSE, digits = 3, ...)

\method{latex}{ep}(object, inline = FALSE, digits = 3, coordinates,
  xchar = "x", ...)

\method{latex}{lp_solution}(object, digits = 3, ...)
}
\arguments{
\item{object}{Object to print in LaTeX.}

\item{...}{Additional arguments.}

\item{inline}{Logical; should the latex be written inline (TRUE), i.e. you have started
math mode yourself, or not inline (FALSE), i.e. an \code{align*}
environment is set up for you.}

\item{digits}{Number of rounding digits}

\item{showName}{Logical; should the name of the matrix be printed?}

\item{asTranspose}{Display as the transpose?}

\item{dispname}{Optional display name.}

\item{coordinates}{An integer value from 1 to the length of any ep. Restricts the
display (so you can throw away non-basic variables).  If missing,
all are displayed.}

\item{xchar}{Character (or string I guess!) to name the solution for display purposes.}
}
\description{
Displaying objects of the appropriate type, typeset in
LaTeX.  For use with knitr/sweave.
}
\examples{
x <- matrix(1:25, 5)
latex(x)
}

